IRA - Percent & Base (Error)

+++ TOSS UPGRADE/Models/FM_GeneralReference/FM_GeneralReference_IRA.cs
@@ -27,6 +27,6 @@ namespace TOSS_UPGRADE.Models.FM_GeneralReference
         public int IRAID { get; set; }
         public decimal IRAPercentageShare { get; set; }
         public int IRAPercent { get; set; }
-        public int IRABase { get; set; }
+        public decimal IRABase { get; set; }
     }
 }
